requirements.txt
telethon
google-generativeai
openai
anthropic
python-dotenv
aiohttp
tzdata